% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_framework_examples.R
\name{create_app_example}
\alias{create_app_example}
\title{create_app_example}
\usage{
create_app_example(run = TRUE)
}
\arguments{
\item{run}{boolean if the shiny app should run or
app with ui and server gets returned}
}
\value{
app or NULL
}
\description{
create_app_example
}
\examples{
\dontrun{
example_create_app()
}
}
