% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_framework.R
\name{subtab}
\alias{subtab}
\title{subtab}
\usage{
subtab(..., name, startExpanded = FALSE, css_files = NULL)
}
\arguments{
\item{...}{tabs, see \code{\link{tab}}}

\item{name}{name to display in left-sidebar for the subtab}

\item{startExpanded}{should it start as expanded?}

\item{css_files}{css file paths}
}
\description{
subtab
}
