% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_framework.R
\name{tab}
\alias{tab}
\title{tab}
\usage{
tab(name, R6class, selected = NULL, css_files = NULL, lazy = TRUE)
}
\arguments{
\item{name}{name to display in left-sidebar for the tab}

\item{R6class}{R6 class with ns, app_rv and logger as arguments and
public ui() and server() functions}

\item{selected}{should the tab be selected on init?}

\item{css_files}{css file paths}

\item{lazy}{should the tabs ui and server only load if a user selects it or
on start-up?}
}
\description{
tab
}
