% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_framework.R
\name{create_app}
\alias{create_app}
\title{Title}
\usage{
create_app(..., app_title = "", css_files_global = NULL)
}
\arguments{
\item{...}{a}

\item{app_title}{a}

\item{css_files_global}{a}
}
\description{
Title
}
